ошибки исправлены

+++ src/main/java/com/homework17/HomeWork17/Service/CalculatorInterface.java
@@ -9,5 +9,5 @@ public interface CalculatorInterface {
 
     int multiply(int num1, int num2);
 
-    float divide(int num1, int num2);
+    double divide(double num1, double num2);
 }
