+++ src/main/java/com/homework17/HomeWork17/Service/CalculatorService.java
@@ -25,11 +25,11 @@ public class CalculatorService implements CalculatorInterface {
     }
 
     @Override
-    public float divide(int num1, int num2) {
+    public double divide(double num1, double num2) {
         if (num2 == 0) {
             throw new ExceptionZeroDivide();
         } else {
-            return (float) num1 / num2;
+            return (double) num1 / num2;
         }
     }
 }
